backend/requirements.txt
playwright>=1.43.0
playwright-stealth>=1.0.6
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
